requirements.txt
anyio==4.6.0
certifi==2024.8.30
charset-normalizer==3.3.2
clickhouse-connect==0.8.0
clickhouse-driver==0.2.9
django-environ==0.11.2
et-xmlfile==1.1.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
lz4==4.3.3
numpy==2.1.1
openpyxl==3.1.5
pandas==2.2.3
PyPDF2==3.0.1
pyTelegramBotAPI==4.23.0
python-dateutil==2.9.0.post0
python-telegram-bot==21.6
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
telebot==0.0.5
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
zstandard==0.23.0
